refactor(core/cms-layout/data-access): #232 use createFeature function for cms-layout state feat
