+++ libs/core/cms-layout/data-access/src/index.ts
@@ -1,6 +1,5 @@
 export * from './lib/+state/layout.actions';
 export * from './lib/+state/layout.facade';
-export * from './lib/+state/layout.selectors';
 export * from './lib/core-cms-layout-data-access.module';
 export * from './lib/core-cms-layout-header-config';
 export * from './lib/core-cms-view-config';
