+++ libs/core/cms-layout/data-access/src/lib/core-cms-layout-data-access.module.ts
@@ -4,13 +4,9 @@ import { StoreModule } from '@ngrx/store';
 import { SharedActivityDataAccessModule } from '@plastik/shared/activity/data-access';
 
 import { LayoutEffects } from './+state/layout.effects';
-import { layoutReducer, LAYOUT_FEATURE_KEY } from './+state/layout.reducer';
+import { selectLayoutFeature } from './+state/layout.feature';
 
 @NgModule({
-  imports: [
-    StoreModule.forFeature(LAYOUT_FEATURE_KEY, layoutReducer),
-    EffectsModule.forFeature([LayoutEffects]),
-    SharedActivityDataAccessModule,
-  ],
+  imports: [StoreModule.forFeature(selectLayoutFeature), EffectsModule.forFeature([LayoutEffects]), SharedActivityDataAccessModule],
 })
 export class CoreCmsLayoutDataAccessModule {}
